perl: Preserve attributes when applying cross files
The file attributes (specifically the timestamps) must be preserved when
applying the cross files to have reproducible builds. Otherwise, the
cross files get the timestamp from when they are copied in the do_unpack
task, and the reproducible_build class uses this timestamp since it is
the youngest one in the source tree.

(From OE-Core rev: 1ddd2c1e3b70744fbcee688482adf08a8f693d45)

Signed-off-by: Joshua Watt <JPEWhacker@gmail.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>